requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
flask_migrate

flask-cors
google-generativeai
vaderSentiment
gunicorn
flask_sqlalchemy